SONARJAVA-997 Define properties to provide test classpath
