+++ sonar-java-plugin/src/test/java/org/sonar/plugins/java/JavaPluginTest.java
@@ -27,7 +27,7 @@ public class JavaPluginTest {
 
   @Test
   public void test() {
-    assertThat(new JavaPlugin().getExtensions().size()).isEqualTo(26);
+    assertThat(new JavaPlugin().getExtensions().size()).isEqualTo(28);
   }
 
 }
